# performance improvement for sampling