--- a/qupulse/hardware/util.py
+++ b/qupulse/hardware/util.py
@@ -95,6 +95,7 @@ def get_sample_times(waveforms: Union[Collection[Waveform], Waveform],
         segment_lengths.append(rounded_segment_length)
 
     segment_lengths = np.asarray(segment_lengths, dtype=np.uint64)
-    time_array = np.arange(np.max(segment_lengths)) / float(sample_rate_in_GHz)
+    float_sample_rate = float(sample_rate_in_GHz)
+    time_array = np.arange(0, np.max(segment_lengths)/float_sample_rate, float_sample_rate) 
 
     return time_array, segment_lengths