This PR improves sampling speed of the `get_sample_times` method.

```
%timeit np.arange(10000)/2.4
21 µs ± 132 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit np.arange(0, 10000/2.4, 1/2.4)
6.81 µs ± 28.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

a=np.arange(10000)/2.4

b=np.arange(0, 10000/2.4, 1/2.4)

a-b
Out[184]: array([0., 0., 0., ..., 0., 0., 0.])

np.max(np.abs(a-b))
Out[185]: 0.0
```

@terrorfisch