@terrorfisch 2021-03-15
Thanks! Unfortunately we have systems which use 2.3GHz where 
```
a=np.arange(10000)/2.3
b=np.arange(0, 10000/2.3, 1/2.3)
np.abs(a-b).max()
Out[9]: 9.094947017729282e-13
```

I guess the expensive thing here are the allocations so one optimization is 
```
time_array = np.arange(np.max(segment_lengths), dtype=float)
time_array /= float(sample_rate_in_GHz)

In [32]: %timeit c = np.arange(10000, dtype=float); c /= 2.4
18.4 µs ± 949 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [33]: %timeit a = np.arange(10000)/2.4
25.6 µs ± 579 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [34]: %timeit b = np.arange(0, 10000/2.4, 1/2.4)
8.08 µs ± 69.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```


This function might be a good candidate for a decoration with `functools.lru_cache` or sample rate based caching. Or do you need the cold start performance?

```python
CACHE = {}

def _get_sample_times(n: int, f: float):
    arr = CACHE.get(f)    
    if arr is None or arr.size < n:
        arr = np.arange(n, dtype=np.float64)
        arr /= f
        CACHE[f] = arr
    elif arr.size > n:
        arr = arr[:n]
    return arr
```