@peendebak 2021-03-15
@terrorfisch I updated the PR to use your variation. Using a cache would add some more performance, but we would have to make the array readonly to make it safe. 
Cold start performance would not be an issue for our use case.